Integration test cleanup
move code from BaseIT to BaseKafkaConnectIT

get rid of protected and use public/private in base classes
