+++ google-cloud-bigtable-kafka-connect-sink/src/test/java/com/google/cloud/kafka/connect/bigtable/integration/BaseKafkaConnectBigtableIT.java
@@ -33,9 +33,10 @@ public abstract class BaseKafkaConnectBigtableIT extends BaseKafkaConnectIT {
   // Not copied from BigtableSinkConfig since it isn't present in its public API.
   public static long DEFAULT_BIGTABLE_RETRY_TIMEOUT_MILLIS = 90000;
 
-  protected BigtableDataClient bigtableData;
-  protected BigtableTableAdminClient bigtableAdmin;
   private final Logger logger = LoggerFactory.getLogger(BaseKafkaConnectBigtableIT.class);
+
+  public BigtableDataClient bigtableData;
+  public BigtableTableAdminClient bigtableAdmin;
 
   @Before
   public void setUpBigtable() {
@@ -55,18 +56,17 @@ public abstract class BaseKafkaConnectBigtableIT extends BaseKafkaConnectIT {
   }
 
   public Map<ByteString, Row> readAllRows(BigtableDataClient bigtable, String table) {
-    Boolean success = null;
+    Integer numRecords = null;
     try {
       Query query = Query.create(table);
       Map<ByteString, Row> result =
           bigtable.readRows(query).stream().collect(Collectors.toMap(Row::getKey, r -> r));
-      success = true;
+      numRecords = result.size();
       return result;
     } catch (Throwable t) {
-      success = false;
       throw t;
     } finally {
-      logger.info("readAllRows({}): success={}", table, success);
+      logger.info("readAllRows({}): #records={}", table, numRecords);
     }
   }
