+++ google-cloud-bigtable-kafka-connect-sink/src/test/java/com/google/cloud/kafka/connect/bigtable/integration/BaseKafkaConnectBigtableSchemaRegistryIT.java
@@ -20,7 +20,7 @@ import org.junit.After;
 import org.junit.Before;
 
 public abstract class BaseKafkaConnectBigtableSchemaRegistryIT extends BaseKafkaConnectBigtableIT {
-  protected SchemaRegistryTestUtil schemaRegistry;
+  public SchemaRegistryTestUtil schemaRegistry;
 
   @Before
   public void setUpSchemaRegistry() throws Exception {
