+++ google-cloud-bigtable-kafka-connect-sink/src/test/java/com/google/cloud/kafka/connect/bigtable/integration/BaseKafkaConnectIT.java
@@ -25,28 +25,50 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Properties;
+import java.util.Set;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
+import org.apache.commons.lang.StringUtils;
+import org.apache.kafka.clients.admin.Admin;
+import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.consumer.ConsumerRecords;
 import org.apache.kafka.clients.producer.KafkaProducer;
+import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.clients.producer.RecordMetadata;
+import org.apache.kafka.common.utils.Utils;
 import org.apache.kafka.connect.data.SchemaAndValue;
 import org.apache.kafka.connect.runtime.SinkConnectorConfig;
+import org.apache.kafka.connect.runtime.WorkerConfig;
 import org.apache.kafka.connect.runtime.errors.DeadLetterQueueReporter;
 import org.apache.kafka.connect.runtime.errors.ToleranceType;
+import org.apache.kafka.connect.runtime.isolation.PluginDiscoveryMode;
 import org.apache.kafka.connect.storage.Converter;
+import org.apache.kafka.connect.util.clusters.EmbeddedConnectCluster;
 import org.junit.After;
 import org.junit.Before;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public abstract class BaseKafkaConnectIT extends BaseIT {
-  private static final long PRODUCE_TIMEOUT_MILLIS = 15000L;
+  public static final long PRODUCE_TIMEOUT_MILLIS = 15000L;
+  // https://cloud.google.com/bigtable/docs/reference/admin/rpc/google.bigtable.admin.v2#createtablerequest
+  public static int MAX_BIGTABLE_TABLE_NAME_LENGTH = 50;
+
+  private final Logger logger = LoggerFactory.getLogger(BaseKafkaConnectIT.class);
+
+  public EmbeddedConnectCluster connect;
+  public Admin kafkaAdminClient;
+  public int numWorkers = 1;
+  public int numBrokers = 1;
 
   @Before
   public void setUpConnect() {
@@ -56,6 +78,99 @@ public abstract class BaseKafkaConnectIT extends BaseIT {
   @After
   public void tearDownConnect() {
     stopConnect();
+  }
+
+  public void startConnect() {
+    logger.info("Starting embedded Kafka Connect cluster...");
+    Map<String, String> workerProps = new HashMap<>();
+    workerProps.put(WorkerConfig.OFFSET_COMMIT_INTERVAL_MS_CONFIG, Long.toString(10000));
+    workerProps.put(WorkerConfig.PLUGIN_DISCOVERY_CONFIG, PluginDiscoveryMode.HYBRID_WARN.name());
+
+    Properties brokerProps = new Properties();
+    brokerProps.put("socket.request.max.bytes", maxKafkaMessageSizeBytes);
+    brokerProps.put("message.max.bytes", maxKafkaMessageSizeBytes);
+    brokerProps.put("auto.create.topics.enable", "false");
+    brokerProps.put("delete.topic.enable", "true");
+
+    Map<String, String> clientConfigs = new HashMap<>();
+    clientConfigs.put(
+        ProducerConfig.MAX_REQUEST_SIZE_CONFIG, String.valueOf(maxKafkaMessageSizeBytes));
+    clientConfigs.put(
+        ProducerConfig.BUFFER_MEMORY_CONFIG, String.valueOf(maxKafkaMessageSizeBytes));
+    clientConfigs.put(
+        ConsumerConfig.MAX_PARTITION_FETCH_BYTES_CONFIG, String.valueOf(maxKafkaMessageSizeBytes));
+    connect =
+        new EmbeddedConnectCluster.Builder()
+            .name("kcbt-connect-cluster-" + getTestClassId())
+            .numWorkers(numWorkers)
+            .numBrokers(numBrokers)
+            .brokerProps(brokerProps)
+            .workerProps(workerProps)
+            .clientConfigs(clientConfigs)
+            .build();
+
+    // Start the clusters
+    connect.start();
+    try {
+      connect
+          .assertions()
+          .assertAtLeastNumWorkersAreUp(1, "Initial group of workers did not start in time.");
+    } catch (InterruptedException e) {
+      throw new RuntimeException(e);
+    }
+    kafkaAdminClient = connect.kafka().createAdminClient();
+    logger.info(
+        "Started embedded Kafka Connect cluster using bootstrap servers: {}",
+        connect.kafka().bootstrapServers());
+  }
+
+  public void stopConnect() {
+    logger.info(
+        "Stopping embedded Kafka Connect cluster using bootstrap servers: {}",
+        connect.kafka().bootstrapServers());
+    if (kafkaAdminClient != null) {
+      Utils.closeQuietly(kafkaAdminClient, "Admin client for embedded Kafka cluster");
+      kafkaAdminClient = null;
+    }
+
+    // Stop all Connect, Kafka and Zk threads.
+    if (connect != null) {
+      Utils.closeQuietly(connect::stop, "Embedded Connect, Kafka, and Zookeeper clusters");
+      connect = null;
+    }
+  }
+
+  public String startSingleTopicConnector(Map<String, String> configProps)
+      throws InterruptedException {
+    return startConnector(configProps, Collections.emptySet());
+  }
+
+  public String startMultipleTopicConnector(
+      Map<String, String> configProps, Set<String> topicNameSuffixes) throws InterruptedException {
+    return startConnector(configProps, topicNameSuffixes);
+  }
+
+  private String startConnector(Map<String, String> configProps, Set<String> topicNameSuffixes)
+      throws InterruptedException {
+    int longestSuffix = topicNameSuffixes.stream().mapToInt(String::length).max().orElse(0);
+    String id =
+        StringUtils.right(
+            getTestCaseId() + System.currentTimeMillis(),
+            MAX_BIGTABLE_TABLE_NAME_LENGTH - longestSuffix);
+    if (topicNameSuffixes.isEmpty()) {
+      configProps.put(SinkConnectorConfig.TOPICS_CONFIG, id);
+      connect.kafka().createTopic(id, numBrokers);
+    } else {
+      configProps.put(SinkConnectorConfig.TOPICS_REGEX_CONFIG, id + ".*");
+      for (String suffix : topicNameSuffixes) {
+        connect.kafka().createTopic(id + suffix, numBrokers);
+      }
+    }
+    connect.configureConnector(id, configProps);
+    connect
+        .assertions()
+        .assertConnectorAndAtLeastNumTasksAreRunning(id, numTasks, "Connector start timeout");
+    return id;
   }
 
   public KafkaProducer<byte[], byte[]> getKafkaProducer() {
